+++ packages/storefront/mdx-components.tsx
@@ -34,7 +34,6 @@ export const P = ({ children }: PropsWithChildren) => (
 export function useMDXComponents(components: MDXComponents): MDXComponents {
   return {
     wrapper: ({ children }) => (
-      // biome-ignore lint/correctness/useUniqueElementIds: <Needed>
       <article id="main-content" className="col-span-full xs:col-start-2 xs:col-end-12">
         {children as React.ReactNode}
       </article>
@@ -107,7 +106,7 @@ export function useMDXComponents(components: MDXComponents): MDXComponents {
       </PLinkPure>
     ),
     blockquote: ({ children }) => (
-      <blockquote className="my-fluid-sm ps-static-md border-s-4 border-solid border-contrast-low">
+      <blockquote className="my-fluid-sm ps-static-md border-s-4 border-solid border-contrast-20">
         {children as React.ReactNode}
       </blockquote>
     ),