+++ packages/storefront/src/components/components/ComponentMetaOverview.tsx
@@ -122,7 +122,7 @@ export const ComponentMetaOverview = () => {
                 <code className="inline-block my-static-xs">
                   {/* biome-ignore lint/a11y/noStaticElementInteractions: ok */}
                   <span
-                    className="hover:text-contrast-medium cursor-pointer"
+                    className="hover:text-contrast-50 cursor-pointer"
                     onClick={(e) => (e.currentTarget.nextElementSibling as HTMLElement).classList.toggle('hidden')}
                   >
                     {propName} {getFlags(propMeta)}